Expose NonceId generation
